Change llvm over to a choice of install or cache
Pass a flag into setup_build llvm_source which dictates whether it should
get the llvm from cache or install. It will set a symbolic link if
required for an install so it's in the same place. All pr tests if run
as a pull request will run using an install except windows where this
is not yet supported.