Switch to calling an action for PR tests and populate the cache sequentially
There was a race condition in updating the cache as it just ran the tests
overnight in parallel.

Move all the build parts of PR jobs to an action. Add a workflow
pr_tests_cache.yml to seed the caches by calling the actions in sequence.
The scheduled run_pr_tests overnight will no longer have the ability to
seed the cache and this will only be done via pr_tests_cache.yml.

This does not yet delete the caches. We can't be sure this works so this
is initially commented out.