Added Xfail and Mayfail attributes to city runner
Support attribute Xfail and Mayfail for city runner. These can be set
as additional elements in the csv file.

Xfail means we expect it to fail and will give a failing return code
if it unexpectedly pass, otherwise it does not contribute to an overall
fail.

Mayfail means it sometimes fails and will not contribute to a failing
return code.

Unlike some of the other attributes, these cannot be set with additional
csv files, as it was felt this leads to overcomplexity. Instead it is
expected that some form of csv combination script could be used alongside
this.