Fix CL_KERNEL_LOCAL_MEM_SIZE a third time.
My previous change made it so that ComputeLocalMemoryUsagePass assuredly
ran before ReplaceLocalModuleScopeVariablesPass which is necessary for
correct results, but also made it so that with deferred compilation, it
could run after encoding the kernel metadata, meaning the correct
results came in too late. Previously, it was getting additionally run in
HostModule::createKernel so that results would always be available.
Restore this.