Change planned testing to build llvm and run all tests.
The current system was very limited to being able to cope with updates
of llvm. It also did not allow us to run all testing on one llvm or the
other. Additionally, it lacked flexibility.

Build llvm as an artefact before testing. Option to support using caching
on a per llvm basis - but unlike before this will build it if it does not
find it.

planned_testing_caller.yml has been changed to call both internal tests
(run_ock_internal_tests.yml renamed from run_pr_tests.yml) and external
tests (run_ock_external_tests.yml, renamed from planned_testing.yml). It
has also been changed so it can only be run via being called. It allows
a wide variety of inputs so different variants of the pipeline can be run.

At the top of the planned testing change are per llvm callers,
planned_testing_caller_19.yml and planned_testing_caller_20.yml. These
do little more than call planned_testing.yml with different llvms on a
cron job. They are not expected to change much. llvm_source is passed
down to the run_* workflows, and has a run id option which means it
downloads it from that run id. If it is the same workflow id as being
run on, then it will use download-artifact, otherwise it will use gh.

This additionally switches the overnight testing to llvm 19 and llvm 20
from a previous 18/19 mix, run on different nights.

Some extra flexibility has also been added for which runners are used as
it was difficult to test with the conflict on available runners.