[AArch64] Fix stub call generation.
If we encountered relocations that did not fit in a direct jump, we
generate a trampoline that does a far jump to the real function. This
trampoline contains four mov* instructions to load the individual bytes
of the address to jump to, but then inexplicably skips over the first
of those mov* instructions.

This is not currently being hit in testing because we happen to load
kernels in memory addresses that do not require trampolines, and even if
we force the use of trampolines unnecessarily by adjusting the below
R_AARCH64_CALL26/R_AARCH64_JUMP26 handling, it happens to be the case
that the high byte already has the correct value.