Update to llvm 19/20 being the supported versions
We are now officially dropping 18. This updates the documentation and
PR testing to match.

The default for run_ock_internal_tests.yml has been dropped. This needs
to be set at the higher level of run_pr_tests_caller.yml or
planned_testing_caller.yml.

Fixed windows job for run_internal_test.yml to not run on a pull request.
Remove create_llvm workflow as this is no longer needed.