Set SPIR-V ByVal and Sret alignment
In SPIR-V, the ByVal and Sret attributes on function parameters are not
stated to specify any requirements on alignment. In LLVM, the byval
attribute on function parameters implies "a target-specific assumption"
on alignment if no explicit alignment is specified, which is an
incorrect translation: we need to ensure that no invalid assumptions are
made, so explicitly set the alignment to 1. Setting an explicit
alignment also works around LLVM bug 138356 where the ARM backend
assumes that alignment is always set.