[LLVM 21] Use llvm::driver::createTLII.
LLVM 21 renames llvm::TargetLibraryInfoImpl::LIBMVEC_X86 to
llvm::TargetLibraryInfoImpl::LIBMVEC and changes it to apply across
targets. But there is also a utility function llvm::driver::createTLII
which has existed for longer, which we were effectively re-implementing,
and which avoids our need to know about this rename. Just call that.