[spirv-ll][NFC] Migrate tests to OpenCL API.
Move many spirv-ll tests from the Vulkan API to the OpenCL API.

Distinct signed and unsigned types do not exist, e.g. both signed and
unsigned i32 are represented as OpTypeInt 32 0. Tests are updated
accordingly, and where this results in identical tests for the formerly
different signed and unsigned variants, the tests are merged.