[CI] Specify specific OCK commit to fetch when building DPC++.
We were fetching refs/pull/* in order to ensure that we would always
have pull/<id>/merge available, without accounting for the fact that
pull/<id>/merge may change after CI has been triggered. But we should
not need to fetch all PRs anyway, we only need the specific commit that
we are trying to build, so explicitly fetch that instead.

Note: This approach cannot be used with arbitrary ock_git_tag. We were
not using that functionality so this removes the option.