Update copy of Google Benchmark.
Our copy of Google Benchmark was last updated in 2021 and was relying on
more outdated libraries. Update to the current version.

The KernelCreateEmptyKernelFromBinary test was testing the performance
of loading a SPIR 1.2 binary. Support for this was removed from oneAPI
Construction Kit in #73 but we never updated BenchCL to match, so do
that now.