[spirv-ll] Use new debug info format.
We were using the old debug info format because we were not set up to
allow spv::Id to be mapped to llvm::DbgRecord, and the old debug info
format used llvm::Instruction for that avoiding the problem. However,
the old debug info format cannot be mixed with the new debug info format
and results in errors as LLVM gets updated to automatically use the new
debug info format in more cases.

This commit changes spirv-ll to track LLVM objects using an
llvm::PointerUnion, allowing llvm::DbgRecord to easily be included. It
is meant to be NFC for valid SPIR-V modules, though in corner cases the
error handling of invalid SPIR-V modules may change.