[LLVM 21] Remove getLLVMContext()
LLVM 21 removes ThreadSafeContext::getContext() in favor of
ThreadSafeContext::withContextDo() which takes care of locking, and
highlights that the locking we were previously doing was wrong:
ThreadSafeContext provided us with a mutex to use to give us exclusive
access to the context, and we did not use it, we used our own instead.
Fixing this requires a refactoring.

* compiler::BaseContext is updated to no longer support locking.
* BaseTarget::getLLVMContext is removed.
* BaseTarget::withLLVMContextDo is added, analogous to LLVM's new
  withContextDo function, along with a convenience overload to allow
  returning values.
* Functions are updated to use BaseTarget::withLLVMContextDo instead of
  std::lock_guard<compiler::BaseContext> according to the general
  principle that functions that take an LLVM context or LLVM module
  directly need to be called with the lock already held, functions that
  do not take an LLVM context or LLVM module need to use
  withLLVMContextDo as appropriate.