[cargo][NFC] Use std::variant in argument parser.
Through argument's Type field and its anonymous union, we were
effectively creating our own custom std::variant, except less featureful
and not capable of handling not-trivially-destructible types, which is
why CustomHandler was kept out of the union. There is no advantage to
this hand-coded version and it results in compiler warnings that then
need to be suppressed. Just use std::variant directly instead.