city runner: do not mark skipped tests as failing.
If we encounter a new test, we previously only allowed three scenarios:
pass, timeout, or fail. If the test did not pass and did not time out,
we treated it as a fail. This is incorrect for tests that self-report as
skipped, which need to be treated as a separate category. This change
implements that.