Revert "Merge pull request #950 from coldav/colin/clang_format_20"
This reverts commit 035a289116a01c39cefd8a6362813743776be115.

This went beyond the clang-format version update that we needed in
changing code style throughout the project to align with "upstream
dpc++" but DPC++ is not upstream, nor has it ever been, nor is it
planned to be. While we do intend to copy some code over to there, style
changes that are only for the benefit of DPC++ do not belong in this
repository. Because everything was tangled together, it isn't feasible
to revert only the problematic parts, we have to revert the whole thing
and do again the parts we want to keep.