metadata: avoid undefined behavior.
We were calling va_end in a helper lambda, but va_end has a requirement
that "Each invocation of the va_start and va_copy macros shall be
matched by a corresponding invocation of the va_end macro in the same
function." It is not valid to delegate the va_end to a helper function,
not even if that helper function is in the form of a lambda.

A slight refactor to create md_vloadf and md_vpushf helper functions
makes this more easily manageable.