Improve ARM processor checks.
When cross-compiling for ARM, the toolchain file will likely set
CMAKE_SYSTEM_PROCESSOR to exactly "arm", but when building for ARM
natively, that is not the case and we may encounter values such as
"armv7a". Rather than enumerating each and every possible value, treat
anything that starts with "arm" other than "arm64" the same.