Split CTS further.
The current strategy of splitting OpenCL CTS in two separate jobs for
RISC-V is not quite sufficient, as both jobs take about six hours and
they regularly cross the threshold to time out. Split this in four
separate jobs instead.

For consistency, the SYCL-CTS action is updated likewise even though
this is no longer run for RISC-V.