Run OpenCL CTS without RVV.
QEMU's emulation of RISC-V with RVV is significantly slower than RISC-V
without RVV and results in OpenCL CTS timing out almost every time. The
timeout is a hard timeout of 6 hours, we cannot increase that, and we
cannot split up the tests further to cut back the time because some
individual tests already need that long. While it would be nice if we
could test RVV, at the moment it is doing more harm than good because it
means it becomes pretty much impossible for our workflow to ever pass.

This commit changes our run of OpenCL CTS to not use RVV, and our
invocation of QEmu to not enable RVV instructions. This reduces the run
time to under four hours, allowing us to continue doing some RISC-V
testing.