Save kernel even after new program builds.
When a kernel is in a queue to be executed, we need to ensure that the
code that is to be executed is preserved as well. Depending on whether
deferred compilation is in use, the code may be managed either by the
deferred kernel, or by the kernel cache. This commit turns both into
shared objects that can have their lifetime extended if needed.